API Rename ExclamationMark icon component to CircleWarning
